Fix the online program bug
Process become unstable because I use thread in the thread, which is
when camera capture a image, my image processing method will be called
and execute by a thread. In my algorithm, there will run 3 thread at the
same time, thus make 3 thread in a thread.

Simply delete the thread code to fix the bug.